requirements.txt
simpy==4.1.1
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
